refactor: share repositoryRelativePath between IO consumers
