+++ Sources/SwiftCPD/IO/GitRefSourceReader.swift
@@ -20,7 +20,7 @@ struct GitRefSourceReader: SourceReader {
     private let runner: GitProcessRunner
 
     func read(file: String) throws -> Data {
-        let relative = try repositoryRelativePath(for: file)
+        let relative = try repositoryRelativePath(for: file, in: repositoryRoot)
         let spec = "\(resolvedSha):\(relative)"
         let result = try runner.run(
             args: ["cat-file", "blob", spec],
@@ -38,37 +38,5 @@ struct GitRefSourceReader: SourceReader {
         }
 
         return result.stdout
-    }
-}
-
-extension GitRefSourceReader {
-
-    private func repositoryRelativePath(for file: String) throws -> String {
-        let absolute = standardize(
-            file.hasPrefix("/") ? file : repositoryRoot + "/" + file
-        )
-        let normalizedRoot = standardize(repositoryRoot)
-        let rootWithSlash = normalizedRoot.hasSuffix("/") ? normalizedRoot : normalizedRoot + "/"
-
-        if absolute == normalizedRoot {
-            return ""
-        }
-
-        if absolute.hasPrefix(rootWithSlash) {
-            return String(absolute.dropFirst(rootWithSlash.count))
-        }
-
-        if !file.hasPrefix("/") {
-            return file
-        }
-
-        throw FileDiscoveryError.pathOutsideRepository(
-            path: file,
-            repositoryRoot: repositoryRoot
-        )
-    }
-
-    private func standardize(_ path: String) -> String {
-        (path as NSString).standardizingPath
     }
 }
